Ptw/add dockerfile (#4)
Adds a dockerfile and some other common infrastructure to make it easier to work with Docker locally. This requires having docker installed locally. I've added a test notebook and a python script that runs all cells in a notebook to make sure that we can run a notebook through Docker.

* Try adding a python shim

to run a jupyter notebook

* Add test notebook

* Add first draft of dockerfile

* First draft of docker-compose

* Add details on running with Docker